[#46] feat : 마이페이지 경력부분 수정 작업 완료
