+++ src/components/MyPage/Career/index.tsx
@@ -10,10 +10,6 @@ interface CareerProps {
 
 const Career = (props: CareerProps) => {
   const myInfo = JSON.stringify(localStorage.getItem('access_token'))
-  const year = [props.year]
-  const totalYear = year.concat(props.year)
-  const sumYear = totalYear.reduce((a, b) => a + b)
-
   const UpdatePage = useCallback(() => {
     router.push('/sign_up/experience/update')
   }, [router])
@@ -21,7 +17,7 @@ const Career = (props: CareerProps) => {
   return (
     <section css={box}>
       <h2>
-        경력 <strong>총 {sumYear}년차</strong>
+        경력 <strong>총 {props.year}년차</strong>
       </h2>
 
       {/* 회사이름 */}
